+++ dask_geomodeling/raster/sources.py
@@ -140,17 +140,24 @@ class MemorySource(RasterBlock):
         return utils.GeoTransform((p, a, 0, q, 0, -d))
 
     def _get_extent(self):
+        if not self.data.size:
+            return
         bbox = self.geo_transform.get_bbox((0, 0), self.data.shape[1:])
         return utils.Extent(bbox, utils.get_sr(self.projection))
 
     @property
     def extent(self):
-        extent_epsg4326 = self._get_extent().transformed(utils.EPSG4326)
-        return extent_epsg4326.bbox
+        extent = self._get_extent()
+        if extent is None:
+            return
+        return extent.transformed(utils.EPSG4326).bbox
 
     @property
     def geometry(self):
-        return self._get_extent().as_geometry()
+        extent = self._get_extent()
+        if extent is None:
+            return
+        return extent.as_geometry()
 
     def __len__(self):
         return self.data.shape[0]